feat: clipboard comp
